requirements.txt
diskcache>=5.6.3
openai>=2.30.0
fastapi>=0.135.3
uvicorn>=0.42.0
httpx>=0.28.1
python-dotenv>=1.2.2
beautifulsoup4>=4.14.3
lxml>=6.0.2
# Optional — only needed with --search firecrawl
firecrawl-py>=4.21.1
